durian-leaf-disease-detection/pyproject.toml
[project]
name = "durian-leaf-disease-detection"
version = "0.1.0"
description = "Deep Learning model for durian leaf disease detection using YOLOv11x"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ultralytics>=8.3.0",
    "roboflow>=1.1.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "pandas>=2.2.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
]

[project.scripts]
durian-detect = "durian_detect.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/durian_detect"]
